requirements.txt
requests
schedule
SQLAlchemy
python-dotenv
fastapi
uvicorn
pyyamlpytest
pytest-asyncio
